+++ app/build.gradle
@@ -41,7 +41,6 @@ dependencies {
     implementation 'androidx.core:core-ktx:1.8.0'
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.5.1"
     implementation 'androidx.appcompat:appcompat:1.5.1'
-    implementation 'com.google.android.material:material:1.7.0'
     implementation"androidx.fragment:fragment-ktx:1.5.5"
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation 'com.google.ar:core:1.30.0'
@@ -53,4 +52,8 @@ dependencies {
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
     implementation 'com.google.code.gson:gson:2.10'
     implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
+    implementation 'com.google.android.material:material:1.8.0'
+    implementation "androidx.navigation:navigation-fragment-ktx:2.5.3"
+    implementation "androidx.navigation:navigation-ui-ktx:2.5.3"
+    implementation "androidx.fragment:fragment-ktx:1.5.6"
 }
