--- app/src/main/java/com/bignerdranch/android/playlistmaker/main/ui/models/NavigationState.kt
@@ -1,7 +1,0 @@
-package com.bignerdranch.android.playlistmaker.main.ui.models
-
-sealed interface NavigationState {
-    object Search : NavigationState
-    object Library : NavigationState
-    object Settings : NavigationState
-}
