+++ app/src/main/java/com/bignerdranch/android/playlistmaker/medialibrary/ui/child_fragment/PlaylistsFragment.kt
@@ -1,0 +1,26 @@
+package com.bignerdranch.android.playlistmaker.medialibrary.ui.child_fragment
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.bignerdranch.android.playlistmaker.databinding.PlaylistsFragmentBinding
+import com.bignerdranch.android.playlistmaker.medialibrary.ui.view_model.PlaylistsViewModel
+import org.koin.androidx.viewmodel.ext.android.viewModel
+
+class PlaylistsFragment : Fragment() {
+
+    private val viewModel by viewModel<PlaylistsViewModel>()
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?,
+    ): View? {
+        val binding = PlaylistsFragmentBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    companion object {
+        fun newInstance() = PlaylistsFragment()
+    }
+}
