+++ app/src/main/res/layout/fragment_search.xml
@@ -1,0 +1,137 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/container"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    android:background="?attr/colorOnPrimary"
+    tools:context="com.bignerdranch.android.playlistmaker.search.ui.fragment.SearchFragment">
+
+    <TextView
+        android:id="@+id/back_to_main"
+        android:background="?attr/colorOnPrimary"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="start"
+        android:text="@string/search_name"
+        android:textColor="?attr/colorOnSecondary"
+        android:fontFamily="@font/yandex_sans_text_medium"
+        android:textSize="@dimen/button_main_size"
+        android:layout_marginTop="14dp"
+        android:layout_marginStart="20dp"
+        android:letterSpacing="-0.01">
+    </TextView>
+        <FrameLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:background="@drawable/layout_corner"
+            android:layout_marginStart="16dp"
+            android:layout_marginEnd="16dp"
+            android:layout_marginTop="@dimen/top_activity_search">
+
+            <EditText
+                android:id="@+id/inputEditText"
+                android:background="@drawable/layout_corner"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:imeOptions="actionDone"
+                android:singleLine="true"
+                android:drawableStart="@drawable/search_icon"
+                android:drawablePadding="8dp"
+                android:drawableTint="?attr/colorOnSurfaceVariant"
+                android:fontFamily="@font/ys_display_regular"
+                android:hint="@string/search_hint"
+                android:inputType="text"
+                android:focusable="true"
+                android:maxLines="1"
+                android:minHeight="36dp"
+                android:paddingStart="12dp"
+                android:paddingEnd="40dp"
+                android:textColor="@color/main_background_color_background"
+                android:textColorHint="?attr/colorOnSurfaceVariant"
+                android:textFontWeight="400"
+                android:textSize="@dimen/button_search_size" />
+
+            <ImageView
+                android:id="@+id/clearIcon"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_gravity="center|end"
+                android:layout_marginEnd="12dp"
+                android:backgroundTint="?attr/colorOnSurfaceVariant"
+                android:src="@drawable/clear_button"
+                android:visibility="gone" />
+        </FrameLayout>
+
+    <TextView
+        android:id="@+id/label_search"
+        android:background="?attr/colorOnPrimary"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        android:layout_marginTop="50dp"
+        android:text="@string/history_label"
+        android:textColor="?attr/colorOnSecondary"
+        android:fontFamily="@font/yandex_sans_text_medium"
+        android:textSize="19sp"
+        android:letterSpacing="-0.01"
+        android:visibility="gone"/>
+
+    <ImageView
+        android:id="@+id/nothing_search_image"
+        android:layout_width="120dp"
+        android:layout_height="120dp"
+        android:layout_marginTop="110dp"
+        android:layout_gravity="center"
+        android:layout_marginHorizontal="120dp"
+        android:backgroundTint="?attr/colorOnSurfaceVariant"
+        android:fontFamily="@font/ys_display_medium"
+        android:visibility="gone"/>
+
+    <TextView
+        android:id="@+id/placeholderMessage"
+        android:layout_width="300dp"
+        android:layout_height="wrap_content"
+        android:layout_marginHorizontal="24dp"
+        android:layout_marginTop="20dp"
+        android:layout_gravity="center"
+        android:textAlignment="center"
+        android:maxLines="4"
+        android:textColor="?attr/colorSecondary"
+        android:textSize="@dimen/problem_text_size"
+        android:visibility="gone"/>
+
+    <ProgressBar
+        android:id="@+id/progressBar"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="148dp"
+        android:layout_gravity="center"
+        android:visibility="gone"
+        android:indeterminate="true"
+        android:indeterminateTint="@color/progress_bar_color"
+        />
+
+    <Button
+        android:id="@+id/update_button"
+        style="@style/SearchActivityButton"
+        app:cornerRadius="54dp"
+        android:text="@string/text_update_button"/>
+
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/recycler_view"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="@dimen/top_activity_search"
+        android:layout_weight="1" />
+
+    <Button
+        android:id="@+id/clear_history"
+        style="@style/SearchActivityButton"
+        app:cornerRadius="54dp"
+        android:text="@string/clear_history_button"
+        android:visibility="gone"/>
+
+</LinearLayout>
