+++ app/src/main/res/layout/track_item.xml
@@ -4,7 +4,7 @@
     android:id="@+id/track_linear_layout"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:background="?attr/colorPrimaryVariant"
+    android:background="?attr/colorOnPrimary"
     android:orientation="horizontal">
