+++ app/src/main/res/menu/bottom_navigation_menu.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        xmlns:app="http://schemas.android.com/apk/res-auto"
+        android:id="@+id/searchFragment"
+        android:icon="@drawable/search_icon"
+        android:title="@string/search"
+        app:itemIconTint="?attr/colorOnSecondary"/>
+
+    <item
+        android:id="@+id/mediaLibraryFragment"
+        android:icon="@drawable/media_library"
+        android:title="@string/media_library" />
+
+    <item
+        android:id="@+id/settingsFragment"
+        android:icon="@drawable/settings"
+        android:title="@string/settings" />
+</menu>
