+++ app/src/main/res/navigation/main_navigation_graph.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/main_navigation_graph"
+    app:startDestination="@id/mediaLibraryFragment">
+
+    <fragment
+        android:id="@+id/searchFragment"
+        android:name="com.bignerdranch.android.playlistmaker.search.ui.fragment.SearchFragment"
+        android:label="SearchFragment"
+        tools:layout="@layout/fragment_search">
+        <action
+            android:id="@+id/action_searchFragment_to_playerActivity"
+            app:destination="@id/playerActivity" />
+    </fragment>
+    <activity
+        android:id="@+id/playerActivity"
+        android:name="com.bignerdranch.android.playlistmaker.player.ui.activity.PlayerActivity"
+        android:label="PlayerActivity"
+        tools:layout="@layout/player"/>
+    <fragment
+        android:id="@+id/settingsFragment"
+        android:name="com.bignerdranch.android.playlistmaker.settings.ui.fragment.SettingsFragment"
+        android:label="SettingsFragment"
+        tools:layout="@layout/fragment_settings"/>
+    <fragment
+        android:id="@+id/mediaLibraryFragment"
+        android:name="com.bignerdranch.android.playlistmaker.medialibrary.ui.root_fragment.MediaLibraryFragment"
+        android:label="MediaLibraryFragment"
+        tools:layout="@layout/fragment_medialibrary"/>
+</navigation>
