+++ app/src/main/res/values-night/colors.xml
@@ -2,4 +2,6 @@
 <resources>
     <color name="color_status_bar_search_activity">#1A1B22</color>
     <color name="status_bar_color_media_library">#1A1B22</color>
+    <color name="selected_color_bottom_nav_item">#3772E7</color>
+    <color name="unselected_color_bottom_nav_item">#FFFFFF</color>
 </resources>
